api/requirements.txt
PyJWT==2.8.0  # Для работы с JWT токенами
SQLAlchemy  # Для работы с базой данных
psycopg2-binary  # PostgreSQL адаптер
Flask  # Веб-фреймворк
Flask-CORS  # Для CORS поддержки
argon2-cffi  # Для хеширования паролей
pycryptodome  # Для шифрования данных 
python-dotenv  # Для работы с переменными окружения
Pillow  # Для генерации изображений с аватарами